+++ BankingAPPConsole_KristianPerez/Program.cs
@@ -48,14 +48,14 @@ namespace oops_task1
                     #region Withdrawl
                     case 3:
                         System.Console.WriteLine("Enter amount to withdrawl");
-                        newAcc.Withdrawl(Convert.ToDouble(Console.ReadLine()));
+                        newAcc.Withdrawl(Convert.ToInt32(Console.ReadLine()));
                         System.Console.WriteLine("Current Balance: {0:C}", newAcc.accBalance);
                     break;
                     #endregion
                     #region Deposit
                     case 4:
                         System.Console.WriteLine("Enter amount to deposit");
-                        newAcc.Deposit(Convert.ToDouble(System.Console.ReadLine()));
+                        newAcc.Deposit(Convert.ToInt32(System.Console.ReadLine()));
                         System.Console.WriteLine("Current Balance: {0:C}", newAcc.accBalance);
                     break;
                     #endregion
